services/feature_engine/requirements.txt
confluent-kafka
python-dotenv
pandas
